feat(positionnement) : EVAL-334 Refactor of code for more logic in where functions are.
