+++ src/main/java/fr/openent/competences/controllers/UtilsController.java
@@ -297,75 +297,4 @@ public class UtilsController extends ControllerHelper {
             });
         });
     }
-
-    /**
-     * Retourne l'activation de la structure du module présences ainsi que l'activation de la récupération des retards/absences du module présence
-     * @param request
-     */
-    @Get("/init/sync/presences")
-    @ApiDoc("Retourne la liste des identifiants des structures de l'utilisateur la récupération des absences/retards du module presences est activée")
-    @SecuredAction(value="", type = ActionType.AUTHENTICATED)
-    public void initRecuperationAbsencesRetardsFromPresences(final HttpServerRequest request){
-        UserUtils.getUserInfos(eb, request, new Handler<UserInfos>() {
-            @Override
-            public void handle(final UserInfos user) {
-                if(user != null && request.params().contains("structureId")){
-                    final String structureId = request.params().get("structureId");
-                    // Récupération de l'état d'activation du module présences de l'établissement
-                    Future<JsonObject> activationFuture = Future.future();
-                    utilsService.getActiveStatePresences(structureId,event -> formate(activationFuture,event));
-
-                    // Récupération de l'état de la récupération des données du modules présences
-                    Future<JsonObject> syncFuture = Future.future();
-                    utilsService.getSyncStatePresences(structureId,event -> formate(syncFuture,event));
-
-                    CompositeFuture.all(syncFuture, activationFuture).setHandler(
-                            event -> {
-                                if(event.failed()){
-                                    String error = event.cause().getMessage();
-                                    log.error("[initRecuperationAbsencesRetardsFromPresences] : " + error);
-                                    badRequest(request, "[initRecuperationAbsencesRetardsFromPresences] : " + error);
-                                } else{
-                                    JsonObject activationState = activationFuture.result();
-                                    JsonObject syncState = syncFuture.result();
-                                    JsonObject result = activationState.mergeIn(syncState);
-                                    Renders.renderJson(request, result);
-                                }
-                            });
-                }else{
-                    badRequest(request);
-                }
-            }
-        });
-    }
-
-
-    /**
-     * Active ou désactive la récupération des absences/retards de presences sur compétences pour une structure donnée
-     * @param request
-     */
-    @Post("/sync/presences")
-    @ApiDoc("Active la récupération des absences/retards de presences sur compétences pour une structure donnée")
-    @SecuredAction(value="", type = ActionType.AUTHENTICATED)
-    public void activateStructureRecuperationAbsencesRetardsFromPresences(final HttpServerRequest request){
-        UserUtils.getUserInfos(eb, request, new Handler<UserInfos>() {
-            @Override
-            public void handle(final UserInfos user) {
-                RequestUtils.bodyToJson(request, new Handler<JsonObject>() {
-                    @Override
-                    public void handle(JsonObject body) {
-                        if(user != null && body.containsKey("structureId") && body.containsKey("state")){
-                            final String structureId = body.getString("structureId");
-                            final Boolean state = body.getBoolean("state");
-                            Handler<Either<String, JsonObject>> handler = defaultResponseHandler(request);
-                            utilsService.activeDeactiveSyncStatePresences(structureId, state, handler);
-                        }else{
-                            badRequest(request);
-                        }
-                    }
-                });
-            }
-        });
-    }
-
 }
