+++ src/main/java/fr/openent/competences/service/StructureOptionsService.java
@@ -21,4 +21,25 @@ public interface StructureOptionsService extends CrudService {
      */
     void getIsAverageSkills (String structureId, Handler<Either<String, JsonObject>> handler);
 
+    /**
+     * change sync state of data from presenecs
+     * @param idStructure id of the structure
+     * @param state state
+     * @param handler response
+     */
+    void activeDeactiveSyncStatePresences(String idStructure, Boolean state, Handler<Either<String, JsonObject>> handler);
+
+    /**
+     * get sync state of data from presenecs
+     * @param idStructure id of the structure
+     * @param handler response
+     */
+    void getSyncStatePresences(String idStructure, Handler<Either<String, JsonObject>> handler);
+
+    /**
+     * get config viescolaire and activation state of presences modules
+     * @param idStructure id of the structure
+     * @param handler response
+     */
+    void getActiveStatePresences (final String idStructure, Handler<Either<String,JsonObject>> handler);
 }
