+++ src/main/java/fr/openent/competences/service/UtilsService.java
@@ -305,28 +305,6 @@ public interface UtilsService {
      */
     void getClassInfo(final String idClass, Handler<Either<String, JsonObject>> handler);
 
-    /**
-     * get config viescolaire and activation state of presences modules
-     * @param idStructure id of the structure
-     * @param handler response
-     */
-    void getActiveStatePresences (final String idStructure, Handler<Either<String,JsonObject>> handler);
-
-    /**
-     * get sync state of data from presenecs
-     * @param idStructure id of the structure
-     * @param handler response
-     */
-    void getSyncStatePresences(String idStructure, Handler<Either<String, JsonObject>> handler);
-
-    /**
-     * change sync state of data from presenecs
-     * @param idStructure id of the structure
-     * @param state state
-     * @param handler response
-     */
-    void activeDeactiveSyncStatePresences(String idStructure, Boolean state, Handler<Either<String, JsonObject>> handler);
-
     void lauchTransition(List<String> structureIds);
 
     void getYearsAndPeriodes(String idStructure, boolean onlyYear, Handler<Either<String,JsonObject>> handler);
